Move MaximizedWidget APIs under preview/enable-maximized-widget
